+++ CHANGELOG.md
@@ -21,6 +21,8 @@ Bug fixes:
   Document cygwin path translation behavior in the FAQ.
 * [BUG #1741](https://github.com/BurntSushi/ripgrep/issues/1741):
   Fix stdin detection when using PowerShell in UNIX environments.
+* [BUG #1866](https://github.com/BurntSushi/ripgrep/issues/1866#issuecomment-841635553):
+  Fix bug when computing column numbers in `--vimgrep` mode.
 
 
 12.1.1 (2020-05-29)